Moved models to new models folder
